refactor: Применит хук usePosts в логике компонента PostList. Поменяет стиль текста вкладок пользователя. Рефакторит код
