+++ src/app/providers/router/AppRouter.tsx
@@ -1,11 +1,11 @@
 import { FC } from "react";
 import { Routes, Route } from "react-router-dom";
-import { PostsPage } from "../../../pages/PostPage/ui/PostsPage";
-import { UserPostsPage } from "../../../pages/UserPostsPage/ui/UserPostsPage";
-import { UserTodosPage } from "../../../pages/UserTodosPage/ui/UserTodosPage";
-import { PostDetailPage } from "../../../pages/PostDetailPage/ui/PostDetailPage";
-import { UserAlbumsPage } from "../../../pages/UserAlbumsPage/ui/UserAlbumsPage";
-import { AlbumPhotosPage } from "../../../pages/AlbumPhotosPage/ui/AlbumPhotosPage";
+import { PostsPage } from "@/pages/PostPage/ui/PostsPage";
+import { UserPostsPage } from "@/pages/UserPostsPage/ui/UserPostsPage";
+import { UserTodosPage } from "@/pages/UserTodosPage/ui/UserTodosPage";
+import { PostDetailPage } from "@/pages/PostDetailPage/ui/PostDetailPage";
+import { UserAlbumsPage } from "@/pages/UserAlbumsPage/ui/UserAlbumsPage";
+import { AlbumPhotosPage } from "@/pages/AlbumPhotosPage/ui/AlbumPhotosPage";
 
 export const AppRouter: FC = () => {
   return (
