+++ src/features/ThemeSwitcher/ui/ThemeSwitcher.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
-import { Button } from "../../../shared/ui/Button/Button";
-import { useTheme } from "../../../shared/lib/theme/useTheme";
+import { Button } from "@/shared/ui/Button/Button";
+import { useTheme } from "@/shared/lib/theme/useTheme";
 
 export const ThemeSwitcher: FC = () => {
     const [theme, toggleTheme] = useTheme();
