+++ src/pages/PostPage/ui/PostsPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { PostList } from "../../../widgets/PostList/PostList";
+import { PostList } from "@/widgets/PostList/PostList";
 
 export const PostsPage: FC = () => {
   return (
