+++ src/widgets/LayoutHeader/Header.tsx
@@ -1,9 +1,9 @@
 import { Link } from "react-router-dom";
 import styles from "./Header.module.css";
 import { FC, useState, Fragment } from "react";
-import { Modal } from "../../shared/ui/Modal/Modal";
-import { Button } from "../../shared/ui/Button/Button";
-import { ThemeSwitcher } from "../../features/ThemeSwitcher/ui/ThemeSwitcher";
+import { Modal } from "@/shared/ui/Modal/Modal";
+import { Button } from "@/shared/ui/Button/Button";
+import { ThemeSwitcher } from "@/features/ThemeSwitcher/ui/ThemeSwitcher";
 
 export const Header: FC = () => {
     const [isModalOpen, setIsModalOpen] = useState(false);
@@ -42,5 +42,5 @@ export const Header: FC = () => {
               </Modal.Footer>
             </Modal>
         </header>
-    )
+    );
 }
