+++ src/widgets/PostList/PostList.tsx
@@ -1,10 +1,9 @@
-import { useGetPostsQuery } from "../../shared/api/api";
-import { Post, PostLengthFilter } from "../../types/api";
-import { PostCard } from "../../entities/post/ui/PostCard";
-import { withLoading } from "../../shared/lib/hoc/withLoading";
-import { FC, useCallback, useState, useMemo, Fragment } from "react";
-import { filterByLength } from "../../features/PostLengthFilter/lib/filterByLength";
-import { PostLengthFilter as PostLengthFilterComponent } from "../../features/PostLengthFilter/ui/PostLengthFilter";
+import { Post } from "@/types/api";
+import { FC, Fragment } from "react";
+import { PostCard } from "@/entities/post/ui/PostCard";
+import { withLoading } from "@/shared/lib/hoc/withLoading";
+import { usePosts } from "@/features/PostList/model/hooks/usePosts";
+import { PostLengthFilter as PostLengthFilterComponent } from "@/features/PostLengthFilter/ui/PostLengthFilter";
 
 const PostListDefault: FC<{posts: Post[]}> = ({ posts }) => {
   return (
@@ -19,21 +18,7 @@ const PostListDefault: FC<{posts: Post[]}> = ({ posts }) => {
 export const PostListWithLoading = withLoading(PostListDefault);
 
 export const PostList: FC = () => {
-const [filter, setFilter] = useState<PostLengthFilter>("все");
-
-  const { data: posts, isLoading, isError, refetch } = useGetPostsQuery();
-
-  const handleRetry = useCallback(() => {
-    refetch();
-  }, [refetch]);
-
-  const filteredPosts = useMemo(() => {
-    return filterByLength(posts  || [], filter);
-  }, [posts, filter]);
-
-  const handleFilterChange = useCallback((newFilter: PostLengthFilter) => {
-    setFilter(newFilter);
-  }, []);
+  const { posts, isLoading, isError, handleRetry, handleFilterChange, filter } = usePosts();
 
     return (
       <Fragment>
@@ -43,10 +28,10 @@ const [filter, setFilter] = useState<PostLengthFilter>("все");
         />
         
         <PostListWithLoading 
-        posts={filteredPosts} 
+        posts={posts} 
         isLoading={isLoading} 
         isError={isError} 
-        onRetry={handleRetry}
+        onRetry={handleRetry}        
         />
       </Fragment>
     );
