+++ src/widgets/UserTabs/ui/UserTabs.module.css
@@ -15,10 +15,10 @@
   }
   
   .tab:hover {
-    background-color: var(--hover-bg, rgba(0, 123, 255, 0.05));
+    background-color: var(--primary-color, rgba(0, 123, 255, 0.05));
   }
   
   .tab.active {
-    color: rgb(213, 213, 213);
+    color: (var(--primary-color, #007bff));
     border-bottom-color: red;
   }
